Rename widgets back to craftcomponents to make review easier
