+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/OpenToParticipation/ProjectCarrousel/LightProjectCard.tsx
