+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/OpenToParticipation/ProjectCarrousel/ProjectCarrousel.stories.tsx
