+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/Spotlight/Settings/index.tsx
