+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/Spotlight/Spotlight.tsx
