+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/_deprecated/Proposals/index.tsx
