+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Settings/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import SettingsWrapper from 'components/admin/ContentBuilder/Settings';
 
-import { WIDGET_TITLES } from '../Widgets';
+import { WIDGET_TITLES } from '../CraftComponents';
 
 const HomepageBuilderSettings = () => {
   return <SettingsWrapper titles={WIDGET_TITLES} />;
