+++ front/app/containers/HomePage/SignedInHeader/CTA.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import useLocalize from 'hooks/useLocalize';
 
-import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/HomepageBanner';
+import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/HomepageBanner';
 
 import BannerButton, {
   BannerButtonStyle,
