+++ front/app/containers/HomePage/SignedOutHeader/HeaderContent.tsx
@@ -6,7 +6,7 @@ import styled from 'styled-components';
 
 import useLocalize from 'hooks/useLocalize';
 
-import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/HomepageBanner';
+import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/HomepageBanner';
 import { triggerAuthenticationFlow } from 'containers/Authentication/events';
 
 import AvatarBubbles from 'components/AvatarBubbles';
