+++ front/app/containers/HomePage/SignedOutHeader/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/HomepageBanner';
+import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/HomepageBanner';
 
 import FixedRatioLayout from './FixedRatioLayout';
 import FullWidthBannerLayout from './FullWidthBannerLayout';
